Move consistency check logic into a checker struct
This makes it possible to get rid of the globals without passing
multiple state vars around.
